% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swampeco-packages.R
\docType{package}
\name{SwampEcoPkgs-package}
\alias{SwampEcoPkgs}
\alias{SwampEcoPkgs-package}
\title{SwampEcoPkgs: Load commonly used R packages}
\description{
Loads CRAN and Github packages commonly used by author and fellow analysts.
}
\author{
\strong{Maintainer}: Paul Julian \email{pauljulianphd@gmail.com}

}
\keyword{internal}
